Revert "Put CallbackManager behind an interface."
