+++ common/src/com/dmdirc/parser/common/BaseParser.java
@@ -69,7 +69,7 @@ public abstract class BaseParser extends ThreadedParser {
     private URI proxy;
 
     /** The callback manager to use for this parser. */
-    private CallbackManager callbackManager;
+    private final CallbackManager callbackManager;
 
     /**
      * Creates a new base parser for the specified URI.
@@ -87,7 +87,7 @@ public abstract class BaseParser extends ThreadedParser {
             }
         }
 
-        callbackManager = new CallbackManagerImpl(implementations);
+        callbackManager = new CallbackManager(implementations);
         callbackManager.initialise(this);
     }
 
@@ -159,11 +159,6 @@ public abstract class BaseParser extends ThreadedParser {
     @Override
     public CallbackManager getCallbackManager() {
         return callbackManager;
-    }
-
-    @Override
-    public void setCallbackManager(final CallbackManager callbackManager) {
-        this.callbackManager = callbackManager;
     }
 
     /**
